requirements.txt
fastapi
uvicorn
pydantic
numpy<2
rdkit-pypi
